requirements.txt
asgiref==3.2.7
Django==3.0.6
django-modeltranslation==0.15
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
